+++ lib/models/api/v2/entities/url_entity_v2.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 
 import 'package:json_annotation/json_annotation.dart';
-import 'package:tweet_ui/models/api/entieties/url_entity.dart';
+import 'package:tweet_ui/models/api/v1/entieties/url_entity.dart';
 import 'package:tweet_ui/models/api/v2/entities/entity_v2.dart';
 
 part 'url_entity_v2.g.dart';