+++ lib/tweet_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:tweet_ui/default_text_styles.dart';
-import 'package:tweet_ui/models/api/tweet.dart';
+import 'package:tweet_ui/models/api/v1/tweet.dart';
 import 'package:tweet_ui/models/api/v2/tweet_v2.dart';
 import 'package:tweet_ui/models/viewmodels/tweet_vm.dart';
 import 'package:tweet_ui/on_tap_image.dart';
@@ -93,8 +93,8 @@ class TweetView extends StatelessWidget {
     this.videoHighQuality,
   }); //  TweetView(this.tweetVM);
 
-  TweetView.fromTweet(
-    Tweet tweet, {
+  TweetView.fromTweetV1(
+    TweetV1Response tweet, {
     this.userNameStyle = defaultUserNameStyle,
     this.userScreenNameStyle = defaultUserScreenNameStyle,
     this.textStyle = defaultTextStyle,