+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update Android SDK versions
 - Update cached_network_image, video_player, better_player, photo_view, url_launcher
 - Added alpha version of support for Twitter API V2
+  - [BREAKING] - Changing Tweet to TweetV1Response, changing fromTweet to fromTweetV1
   - Adding dependencies for: build_runner, json_serializable and json_annotation
   - Added new DTOs for Twitter V2 objects
   - Adding converter from Twitter V2 object to TweetVM